Created Generate Test Data Files | modified CMakeList to include new files | Make the code to Adapt to Double Datatype from int

+++ include/item.hpp
@@ -2,5 +2,5 @@
 #include <vector>
  
 struct Item {
-    std::vector<int> coords; // coords[0]..coords[d-1]
+    std::vector<double> coords; // coords[0]..coords[d-1]
 };
