+++ src/threshold.cpp
@@ -1,11 +1,11 @@
 #include "threshold.hpp"
 
-int ThresholdFromV1(
+double ThresholdFromV1(
     const std::vector<Item>& V1,
     size_t d
 )
 {
-    int xmax = V1[0].coords[d-1];
+    double xmax = V1[0].coords[d-1];
 
     for(const auto& p : V1)
     {
